backend/mcp-python-agent/test.py
import os
import dotenv
import google.generativeai as genai

dotenv.load_dotenv()
genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))


model = genai.GenerativeModel("gemini-2.0-flash")

response = model.generate_content("Tell me a joke.")
print(response.text)
